Introduce Caching and Non-caching observables
In order to be able to track dependencies between nodes, also in the
event when no external `UUID` is available, had to allow to have two types
of Observables. Similarly, `UUID`s that are external and internal are now
easily distinguishable thanks to `RuntimeID` wrapper.

Currently some tracking logic is inside nodes themselves. That's
obviously wrong - it should all be in the instrumentation. Follow up
change will move the logic. The current approach is mostly for
achieving a minimal viable proof that the approach will work.