Down to 2 failures in RuntimeServerTests
Mostly cosmetic changes due to the fact that we now invalidate less.

But two prominent changes are included:
- invalidation logic when Recompute requests comes
- Fixed tracking method calls/closure execution

The latter uses enter/exit logic in `RuntimeAnalysis` in
`InvokeCallableNode` which is correct but undeseriable. The information
should be gathered in instrumention instead. Explicit calls are easier
to experiment with and will be transfered to instrumentation eventually.