All RuntimeServerTest cases pass
Had to ensure that dependencies are being tracked correctly across
function call boundaries. This is problematic due to the fact that local
calls are being done with their own RuntimeCaches; any invalidation has
to ensure that it is capable of crossing that boundary.

Additionally, had to ensure that self argument is properly
tracked/invalidated or changes withing the functions would not be
translated into Truffle nodes.

The logic is still experiemental and elements of RuntimeAnalysis will
need to move to instrumentation, for performance reasons.