Enable broken test by testing a workaround
When a text edit is performed on a code that is being called in the
visualization, a full re-evaluation is needed in order to generate
updated Truffle nodes.

Decided not to support that case, as it is very unlikely to be needed in
the foresable future. Text change will invalidate necessary observers,
as one would expect, but it will not trigger re-evaluation. Instead, one
has to explicitly make `ModifyVisualization` request.
This seems like a good compromise as evaluations of visualization
expressions are very costly (also due to locking).