Fix a broken assumption with local variables access
With runtime tracking we are able to precisely discover which cached
values need to be invalidated. Unfortunately excessive use of external
UUIDs (and therefore caching) by GUI revealed a flaw in that logic wrt
to local variable access.

Whenever a local variable is being accessed it is reading a previously
written value from a frame. But if the assignment has been cached, then
a specific slot in the frame is not written during successive execution.
This means that adding new expressions that make use of previously
cached local variables would report uninitialized value errors.

This change workarounds this by **never** caching assignments.
Assignments should be fast if the RHS is already cached, and GUI appears
to assign external UUIDS to RHS always, it seems. That way any reading
from a frame's slot will succeed.

Also reverted to the old tree building in Changeset as it appeared to be
sufficient for our needs.