Fix RuntimeErrorsTest
Old change disabled caching of DataflowErros. That does not appear to be
necessary anymore.
Also disovered a current limitation of the approach when an edit is
being made in a local function that is different from the current
execution context. In that case we are not able to properly
track/invalidate dependencies thus leading to stale computations.
This is hypothetical scenario as GUI would never make such a call.