Ignore MacOS' auxilary directories
It appears that on some MacOS CI runners some indexing is turned on,
which means that OS creates automatically hidden directories. This
messes up with our expectations. This change adds an exception for MacOS
to simply ignore `._` prefixed directories during unpacking in
`Archive.scala`.

The rest of the changes is just to move `Platform` logic to
`os-environment` package where it should be in the first place.