Re-introduce old visualization logic
Rather than associating visualizations with either RuntimeCache or Ref
entries, fallback to the old logic using synchronization. This is due to
the fact on how instrumentation collects node information when entering
some non-top context. It was impossible to track when to send
visualization updates when popping context without a separate
synchronization map which now finally made sense.
5 tests are failing, mostly due to subexpression invalidation (or rather
lack of it). This will be introduced in the next PR.

RuntimeAnalysis should be somehow accessible via state but that is
currently problematic. This will be resolved eventually.