Ensure loggers are setup correct in LS
During the project-manager we broke the logging infrastructure for
simultaneously logging to console, logs and remote endpoints. This is
because it relied on setup that is no longer present.

This change ensures that we do not initializze socket appenders multiple
times, that minimal logging level to file is actually respected and no
duplciate log files are being created.