Track self arg as a dependent of application
Self arguments are cached. This means that if the type definition or its
methos change, then any method invocation on that type also needs to be
invalidated. This is now possible, as invalidating the whole application
will also invalidate the self argument and ensure that types module scope
is updated.

... at the cost of a small number of extra invalidations. But whatever
hack we would add, it would be inevitable.